Language.Drasil.Debug.Print

Main Function

printAllDebugInfo :: PrintingInformation -> [Doc]

Gathers all printing functions and creates the debugging tables from them.

Helpers

Separators

cdbSection :: Doc -> Doc

Debugging table separator.

header :: Doc -> Doc

Header for debugging tables.

Table Generators

mkTableFromLenses :: HasUID a => PrintingInformation -> (ChunkDB -> UMap a) -> String -> [PrintingInformation -> (String, a -> Doc)] -> Doc

General function to make the debugging tables. Takes in printing information, a function that extracts a certain field from the printing information, a title, three column headers, and three functions that sort the data from the printing information field into the required display formats (often UIDs, terms, shortnames, definitions, etc.).

openTerm :: NamedIdea a => PrintingInformation -> (String, a -> Doc)

openSymbol :: HasSymbol a => PrintingInformation -> (String, a -> Doc)

openDefSymbol :: DefinesQuantity s => PrintingInformation -> (String, s -> Doc)

openAbbreviation :: Idea a => PrintingInformation -> (String, a -> Doc)

openDefinition :: Definition a => PrintingInformation -> (String, a -> Doc)

openUnitSymbol :: HasUnitSymbol a => PrintingInformation -> (String, a -> Doc)

openShortName :: HasShortName a => PrintingInformation -> (String, a -> Doc)

openTitle :: PrintingInformation -> (String, Section -> Doc)

cntntToStr :: RawContent -> String

openContentType :: HasContents s => p -> (String, s -> Doc)

openRef :: HasRefAddress a => p -> (String, a -> Doc)

mkTableSymb :: PrintingInformation -> Doc

Makes a table with all symbolic quantities in the SRS.

mkTableOfTerms :: PrintingInformation -> Doc

Makes a table with terms in the SRS.

mkTableConcepts :: PrintingInformation -> Doc

Makes a table with all concepts in the SRS.

mkTableUnitDefn :: PrintingInformation -> Doc

Makes a table with all units used in the SRS.

mkTableDataDef :: PrintingInformation -> Doc

Makes a table with all data definitions in the SRS.

mkTableGenDef :: PrintingInformation -> Doc

Makes a table with all general definitions in the SRS.

mkTableTMod :: PrintingInformation -> Doc

Makes a table with all theoretical models in the SRS.

mkTableIMod :: PrintingInformation -> Doc

Makes a table with all instance models in the SRS.

mkTableCI :: PrintingInformation -> Doc

Makes a table with all concept instances in the SRS.

mkTableLC :: PrintingInformation -> Doc

Makes a table with all labelled content in the SRS.

mkTableRef :: PrintingInformation -> Doc

Makes a table with all references in the SRS.

mkTableDepChunks :: PrintingInformation -> Doc

Chunks that depend on other chunks. An empty list means the chunks do not depend on anything.

mkTableReferencedChunks :: PrintingInformation -> Doc

Chunks that are referenced and used by other chunks. Those chunks build on top of the ones listed here.

mkTableDepReffedChunks :: PrintingInformation -> Doc

Chunks that use and are used by other chunks.

UID Manipulation

renderUsedUIDs :: [(UID, String)] -> Doc

Creates a table of all UIDs and their "highest" recorded level of information. See mkListShowUsedUIDs for more details.

mkListShowUsedUIDs :: PrintingInformation -> [(UID, String)]

For the last section of the log output. Shows which chunk UID is being used at which stage. Note that chunks used at a "higher stage" (like Concepts and QuantityDicts) will still be built off of the more basic types (like IdeaDicts), they are just not explicitly used in that manner. Also, some chunks may have been "downgraded" when put into the database (for example, mapping a QuantityDict wrapper onto things like Constrained and Unital chunks happens often).

mkListAll :: ChunkDB -> [UID]

Get all UIDs from a database (ChunkDB).