Language.Drasil.HTML.Helpers

Documentation

html :: Doc -> Doc

HTML tag wrapper

headTag :: Doc -> Doc

Head tag wrapper

body :: Doc -> Doc

Body tag wrapper

title :: Doc -> Doc

Title tag wrapper

paragraph :: Doc -> Doc

Paragraph tag wrapper

code :: Doc -> Doc

Code tag wrapper

tr :: Doc -> Doc

Table row tag wrapper

th :: Doc -> Doc

Table header tag wrapper

td :: Doc -> Doc

Table cell tag wrapper

figure :: Doc -> Doc

Figure tag wrapper

figcaption :: Doc -> Doc

Figcaption tag wrapper

li :: Doc -> Doc

List tag wrapper

pa :: Doc -> Doc

Paragraph in list tag wrapper

ba :: Doc -> Doc

ol :: [String] -> Doc -> Doc

Ordered list tag wrapper

ul :: [String] -> Doc -> Doc

Unordered list tag wrapper

table :: [String] -> Doc -> Doc

Table tag wrapper

img :: [(String, Doc)] -> Doc

Image tag wrapper

h :: Int -> Doc -> Doc

Helper for HTML headers

data Variation

Constructors

Class

Id

wrap :: String -> [String] -> Doc -> Doc

wrap' :: String -> [String] -> Doc -> Doc

wrapGen' :: ([Doc] -> Doc) -> Variation -> String -> Doc -> [String] -> Doc -> Doc

Helper for wrapping HTML tags. The forth argument provides class names for the CSS.

wrapGen :: Variation -> String -> Doc -> [String] -> Doc -> Doc

wrapInside :: String -> [(String, Doc)] -> Doc

Helper for wrapping attributes in a tag. | The first argument is tag name. | The String in the pair is the attribute name, | The Doc is the value for different attributes.

caption :: Doc -> Doc

Helper for setting up captions

refwrap :: Doc -> Doc -> Doc

reflink :: String -> Doc -> Doc

Helper for setting up links to references

reflinkInfo :: String -> Doc -> Doc -> Doc

Helper for setting up links to references with additional information

reflinkURI :: String -> Doc -> Doc

Helper for setting up links to external URIs

image :: Doc -> Doc -> MaxWidthPercent -> Doc

Helper for setting up figures

em :: Doc -> Doc

Emphasis (italics) tag

sup :: Doc -> Doc

Superscript tag

sub :: Doc -> Doc

Subscript tag

bold :: Doc -> Doc

Bold tag

articleTitle :: Doc -> Doc

Title header

author :: Doc -> Doc

Author header

divTag :: [String] -> Doc -> Doc

Div tag wrapper

spanTag :: [String] -> Doc -> Doc

indent :: Doc -> Doc

makeCases :: [(Expr, Expr)] -> (Expr -> Doc) -> Doc

Create and markup fractions fraction :: Doc -> Doc -> Doc fraction a b = divTag ["fraction"] (spanTag ["fup"] a $$ spanTag ["fdn"] b)

Build case expressions