Language.Drasil.HTML.Print

Documentation

genHTML :: PrintingInformation -> String -> Document -> Doc

Generate an HTML document from a Drasil Document.

renderCite :: Citation -> (Doc, Doc)

For when we add other things to reference like website, newspaper

data OpenClose

Referring to fence (for parenthesis and brackeds). Either opened or closed.

Constructors

Open

Close

fence :: OpenClose -> Fence -> String

Allows for open/closed variants of parenthesis, curly brackets, absolute value symbols, and normal symbols.