Language.Drasil.Markdown.Config

Description

Defines functions to create .toml and .csv config files for mdBook.

Documentation

makeBook :: Document -> PrintingInformation -> Doc

Prints the .toml config file for mdBook.

makeRequirements :: PrintingInformation -> Doc

Prints the .csv file mapping the original filepaths of assets to the location mdBook uses.

mkTitle :: PrintingInformation -> Sentence -> Doc

Render a title Sentence.

assetMat :: PrintingInformation -> [[Filepath]]

Map the original filepaths of assets to the location mdBook generator needs.