Language.Drasil.Markdown.CreateMd

Description

Markdown file creator for generated GOOL code.

Main Function

makeMd :: [Doc] -> Doc

Combines a list of sentences into a final Doc, also appends end note.

Section Creators

introInfo :: String -> [String] -> Maybe String -> Maybe String -> Doc

Example title, authors, and maybe purpose section.

whatInfo :: Maybe String -> Maybe String -> Doc

What section in generated README file, displays description and scope if not empty

verInfo :: String -> String -> Doc

Language version section.

unsupOS :: Maybe String -> Doc

Invalid Operating Systems section, does not display unless atleast 1 invalid OS.

extLibSec :: [(String, String)] -> [String] -> Doc

External Libraries section. The inputs are a list of name and version pairs and a list of the corresponding version numbers, these are first combined into a list of triplets, and then each printed on a new line.

regularSec :: Doc -> Doc -> Doc

Constructs regular section section from header and message.

instDoc :: [String] -> String -> (String, String) -> Doc

Instruction section, contains 4 paragraphs, Running, Building, Input-Output and Config Files. The Config file section is only displayed if there are configuration files.

endNote :: Int -> [String] -> Doc

End section.