Language.Drasil.Markdown.Helpers

Description

Defines helper functions for creating Markdown files.

Documentation

ang :: Doc -> Doc

Angled brackets

bold :: Doc -> Doc

Bold text

em :: Doc -> Doc

Italicized text

li :: Doc -> Doc

List tag wrapper

ul :: Doc -> Doc

Unordered list tag wrapper.

divTag :: Doc -> Doc

Helper for setting up section div

centeredDiv :: Doc -> Doc

Helper for setting up centered div tags

centeredDivId :: Doc -> Doc -> Doc

Helper for setting up centered div tags with an Id

reflink :: RefMap -> String -> Doc -> Doc

Helper for setting up links to references

reflinkInfo :: RefMap -> String -> Doc -> Doc -> Doc

Helper for setting up links to references with additional information.

reflinkURI :: Doc -> Doc -> Doc

Helper for setting up links to external URIs

image :: Doc -> Maybe Doc -> Doc

Helper for setting up figures

caption :: Doc -> Doc

Helper for setting up captions

heading :: Doc -> Doc -> Doc

Helper for setting up headings with an id attribute. id attribute will only work for mdBook.

h :: Int -> Doc

Helper for setting up heading weights in mdBook.

h' :: Int -> Doc

Helper for setting up heading weights in normal Markdown.

docLength :: Doc -> Int

Helper for getting length of a Doc