Language.Drasil.Printers

HTML

Printer

genHTML :: String -> Document -> Doc

Generate an HTML document from a Drasil Document.

Helpers

makeCSS :: Document -> Doc

Generates the CSS selectors necessary for a document.

Markdown

Printer

makeMd :: [Doc] -> Doc

Combines a list of sentences into a final Doc, also appends end note.

Section Printers

introInfo :: String -> [String] -> Maybe String -> Maybe String -> Doc

Example title, authors, and maybe purpose section.

verInfo :: String -> String -> Doc

Language version section.

unsupOS :: Maybe String -> Doc

Invalid Operating Systems section, does not display unless atleast 1 invalid OS.

regularSec :: Doc -> Doc -> Doc

Constructs regular section section from header and message.

instDoc :: [String] -> String -> (String, String) -> Doc

Instruction section, contains 4 paragraphs, Running, Building, Input-Output and Config Files. The Config file section is only displayed if there are configuration files.

extLibSec :: [(String, String)] -> [String] -> Doc

External Libraries section. The inputs are a list of name and version pairs and a list of the corresponding version numbers, these are first combined into a list of triplets, and then each printed on a new line.

endNote :: Int -> [String] -> Doc

End section.

whatInfo :: Maybe String -> Maybe String -> Doc

What section in generated README file, displays description and scope if not empty

Plain

Types

data SingleLine

Data is either linear or not.

Constructors

OneLine

MultiLine

Functions

sentenceDoc :: SingleLine -> Spec -> Doc

Create sentences for a document in Doc format.

exprDoc :: SingleLine -> Expr -> Doc

Create expressions for a document in Doc format.

codeExprDoc :: SingleLine -> Expr -> Doc

Create code expressions for a document in Doc format. assumes someone has already makde the code expression into an expression

symbolDoc :: Symbol -> Doc

Create symbols for a document in Doc format.

unitDoc :: SingleLine -> USymb -> Doc

Helper for printing units in Doc format.

showSymb :: Symbol -> String

Helper for printing Symbols

showHasSymbImpl :: HasSymbol x => x -> String

Helper for printing a HasSymbol in Implementation Stage

TeX

genTeX :: Document -> ShowTableOfContents -> PrintingInformation -> Doc

Generates a LaTeX document.

Jupyter

genJupyterLessonPlan :: Document -> Doc

Build the general Jupyter Notebook document.

genJupyterSRS :: Document -> Doc

Build an SRS document in JSON format.

Markdown

genMDBook :: Project -> [(Filename, Doc)]

Generate a mdBook SRS

makeBook :: Document -> PrintingInformation -> Doc

Prints the .toml config file for mdBook.

makeRequirements :: PrintingInformation -> Doc

Prints the .csv file mapping the original filepaths of assets to the location mdBook uses.

Log

printAllDebugInfo :: PrintingInformation -> UIDMap -> UIDMap -> [Doc]

Gathers all printing functions and creates the debugging tables from them.

Printing Information and Options

data PrintingInformation

Printing information contains a database, a stage, and a printing configuration.

piSys :: ChunkDB -> Map UID Reference -> Stage -> PrintingConfiguration -> PrintingInformation

Builds a document's printing information based on the system information.

class HasPrintingOptions c where

Able to be printed.

Methods

getSetting :: Lens' c Notation

Holds the printing notation.

data Notation

Notation can be scientific or for engineering.

Constructors

Scientific

Engineering

defaultConfiguration :: PrintingConfiguration

Default configuration is for engineering.

plainConfiguration :: PrintingConfiguration

Simple printing configuration is scientific.