Language.Drasil.Printing.Helpers

Description

General helper functions for printing Drasil documents.

Documentation

($^$) :: Doc -> Doc -> Doc infixl 5

Custom infix operator for concatenating two Docs vertically with an empty line in between.

vsep :: [Doc] -> Doc

Concatenate a list of Docs vertically with an empty line in between.

bslash :: Doc

Single backslash.

Basic text-rendering helper function.

dbs :: Doc

Double backslash.

Basic text-rendering helper function.

assign :: Doc

Variable assignment character ("=").

Basic text-rendering helper function.

eq :: Doc

Equality character ("==").

Basic text-rendering helper function.

lt :: Doc

Less than.

Basic text-rendering helper function.

gt :: Doc

Greater than.

Basic text-rendering helper function.

leq :: Doc

Less than or equal to.

Basic text-rendering helper function.

geq :: Doc

Greater than or equal to.

Basic text-rendering helper function.

dlr :: Doc

Dollar sign.

Basic text-rendering helper function.

ast :: Doc

Asterisk.

Basic text-rendering helper function.

pls :: Doc

Plus.

Basic text-rendering helper function.

hat :: Doc

Hat symbol ("^").

Basic text-rendering helper function.

slash :: Doc

Forward slash.

Basic text-rendering helper function.

hyph :: Doc

Hyphen.

Basic text-rendering helper function.

unders :: Doc

Underscore.

Basic text-rendering helper function.

pipe :: Doc

Pipe.

Basic text-rendering helper function.

emptyline :: Doc

Empty line.

Basic text-rendering helper function.

sq :: String -> Doc

Square brackets.

Text-rendering helper for wrapping strings with brackets/braces.

br :: String -> Doc

Curly braces.

Text-rendering helper for wrapping strings with brackets/braces.

dot :: Doc -> Doc

Dot symbol (".")

Text-rendering helper for appending a period/decimal point (dot symbol) or a comma.

comm :: Doc -> Doc

Comma (",")

Text-rendering helper for appending a period/decimal point (dot symbol) or a comma.

dollarDoc :: Doc -> Doc

For wrapping $ on both sides of a Doc.

paren :: String -> String

Wraps in parenthesis.

Basic plaintext (String) wrapping.

brace :: String -> String

Wraps in curly braces.

Basic plaintext (String) wrapping.

dollar :: String -> String

Wraps in dollar signs.

Basic plaintext (String) wrapping.

sqbrac :: String -> String

Wraps in square brackets.

Basic plaintext (String) wrapping.

angbrac :: String -> String

Wraps in angular brackets ("<>").

Basic plaintext (String) wrapping.

upcase :: String -> Doc

Capitalize first letter of string.

Format strings and convert to Doc.

lowcase :: String -> Doc

Make first letter lowercase.

Format strings and convert to Doc.

sufx :: Int -> String

Appends a suffix for a number. Used only on single digit Ints.

sufxer :: Int -> String

Similar to sufx but used on any sized Int.

sufxPrint :: [CiteField] -> String