Language.Drasil.Printing.Import

Description

Re-export top-level functions for easy use across drasil-printers.

Documentation

expr :: Expr -> PrintingInformation -> Expr

Translate Exprs to printable layout AST.

codeExpr :: PrintingInformation -> CodeExpr -> Expr

Translate CodeExprs to printable layout AST Exprs.

literal :: Literal -> PrintingInformation -> Expr

modelExpr :: ModelExpr -> PrintingInformation -> Expr

Translate Exprs to printable layout AST.

space :: PrintingInformation -> Space -> Expr

Render a Space.

spec :: PrintingInformation -> Sentence -> Spec

Translates Sentence to the printable representation of a Sentence (Spec).

symbol :: Symbol -> Expr

Converts a symbol into an expression.

makeDocument :: PrintingInformation -> Document -> Document

Translates from Document to a printable representation of Document.

makeProject :: PrintingInformation -> Document -> Project

Translates from Document to a printable representation of Project.