Language.Drasil.Printing.Import.Document

Description

Defines functions to transform Drasil-based documents into a printable version.

Main Function

makeDocument :: PrintingInformation -> Document -> Document

Translates from Document to a printable representation of Document.

makeProject :: PrintingInformation -> Document -> Project

Translates from Document to a printable representation of Project.

Helpers

createFiles :: PrintingInformation -> [Section] -> [File]

Helper function for creating sections as Files.

createRefMap' :: File -> [(String, Filename)]

Helper function for creating a RefMap for a Document.

createRefMap :: Filename -> LayoutObj -> [(String, Filename)]

Helper function for creating a RefMap for a LayoutObj

createRef :: Filename -> Label -> [(String, Filename)]

Helper function for mapping a Label to a Filename

createLayout :: PrintingInformation -> [Section] -> [LayoutObj]

Helper function for creating sections as layout objects.

createLayout' :: PrintingInformation -> [Section] -> [LayoutObj]

extractSubS :: Int -> Section -> [(Depth, Section)]

Helper for extracting subsections into their own sections.

file :: PrintingInformation -> (Depth, Section) -> File

Helper for converting a Section to a File

sec :: PrintingInformation -> Int -> Section -> LayoutObj

Helper function for creating sections at the appropriate depth.

cel :: PrintingInformation -> Int -> Section -> LayoutObj

layout :: PrintingInformation -> Int -> SecCons -> LayoutObj

Helper for translating sections into a printable representation of layout objects (LayoutObj).

lay :: PrintingInformation -> Contents -> LayoutObj

Helper that translates Contents to a printable representation of LayoutObj. Called internally by layout.

layLabelled :: PrintingInformation -> LabelledContent -> LayoutObj

Helper that translates LabelledContents to a printable representation of LayoutObj. Called internally by lay.

layUnlabelled :: PrintingInformation -> RawContent -> LayoutObj

Helper that translates RawContents to a printable representation of LayoutObj. Called internally by lay.

layCite :: Citation -> Citation

For importing a bibliography.

layField :: CiteField -> CiteField

Helper for translating Citefields into a printable representation of CiteFields

makeL :: PrintingInformation -> ListType -> ListType

Translates lists to be printable.

item :: PrintingInformation -> ItemType -> ItemType

Helper for translating list items to be printable.