Language.Drasil.Printing.Import.Helpers

Documentation

parens :: Expr -> Expr

Helper for inserting parentheses.

digitsProcess :: [Integer] -> Int -> Int -> Integer -> [Expr]

Processes the digits from the floatToDigits function, decimal point position, a counter, and exponent.

processExpo :: Int -> (Int, Int)

Takes the exponent and the Int of the base and gives the decimal point position and processed exponent. This function supports transferring scientific notation to engineering notation. References for standard of Engineering Notation:

https://www.khanacademy.org/science/electrical-engineering/introduction-to-ee/ intro-to-eeaee-numbers-in-electrical-engineering

https://www.calculatorsoup.com/calculators/math/scientific-notation-converter.php

https://en.wikipedia.org/wiki/Scientific_notation

lookupC :: Stage -> ChunkDB -> UID -> Symbol

Given the stage of the symbol, looks up a character/symbol inside a chunk database that matches the given UID.

lookupT :: ChunkDB -> UID -> Sentence

Look up a term given a chunk database and a UID associated with the term.

lookupS :: ChunkDB -> UID -> Sentence

Look up the acronym/abbreviation of a term. Otherwise returns the singular form of a term. Takes a chunk database and a UID associated with the term.

lookupP :: ChunkDB -> UID -> Sentence

Look up the plural form of a term given a chunk database and a UID associated with the term.