Language.Drasil.Printing.Import.ModelExpr

Description

Defines functions to render CodeExprs as printable Exprs.

Documentation

modelExpr' :: PrintingInformation -> Int -> ModelExpr -> Expr

Helper that adds parenthesis to a display expression where appropriate.

mkCall :: PrintingInformation -> Ops -> ModelExpr -> Expr

Helper that creates an expression row given printing information, an operator, and an expression.

mkBOp :: PrintingInformation -> Ops -> ModelExpr -> ModelExpr -> Expr

Helper that creates a binary expression row given printing information, an operator, and two expressions.

neg' :: ModelExpr -> Bool

Helper for properly rendering negation of expressions.

neg :: PrintingInformation -> ModelExpr -> Expr

Render negated expressions.

indx :: PrintingInformation -> ModelExpr -> ModelExpr -> Expr

For printing indexes.

call :: PrintingInformation -> UID -> [ModelExpr] -> Expr

For printing expressions that call something.

eopAdds :: PrintingInformation -> DomainDesc t ModelExpr ModelExpr -> ModelExpr -> Expr

Helper function for addition EOperators.

eopMuls :: PrintingInformation -> DomainDesc t ModelExpr ModelExpr -> ModelExpr -> Expr

Helper function for multiplicative EOperators.

eop :: PrintingInformation -> AssocArithOper -> DomainDesc t ModelExpr ModelExpr -> ModelExpr -> Expr

Helper function for translating EOperators.

sup :: Integer -> [Expr]

Helper function for display nth derivative

modelExpr :: ModelExpr -> PrintingInformation -> Expr

Translate Exprs to printable layout AST.

assocExpr :: Ops -> Int -> [ModelExpr] -> PrintingInformation -> Expr

Common method of converting associative operations into printable layout AST.

setExpr :: Ops -> Int -> [ModelExpr] -> PrintingInformation -> Expr

addExpr :: [ModelExpr] -> AssocArithOper -> PrintingInformation -> [Expr]

Add add symbol only when the second Expr is not negation

mulExpr :: [ModelExpr] -> AssocArithOper -> PrintingInformation -> [Expr]

Helper for rendering printable expressions.

withParens :: PrintingInformation -> ModelExpr -> ModelExpr -> Expr

Helper that adds parenthesis to the first expression. The second expression is written as a superscript attached to the first.

pow :: PrintingInformation -> ModelExpr -> ModelExpr -> Expr

Helper for properly rendering exponents.

renderRealInt :: PrintingInformation -> Symbol -> RealInterval ModelExpr ModelExpr -> Expr

Print a RealInterval.