Language.Drasil.Printing.LayoutObj

Description

Defines types similar to content types in Language.Drasil but better suited for printing.

Documentation

data Document

A document must contain a title, author, and contents (as LayoutObjs).

Constructors

Document Title Author [LayoutObj]

data Project

A Project must contain a title, author, RefMap, and Files.

Constructors

Project Title Author RefMap [File]

data File

A File must contain a title, filename, depth, and contents (as LayoutObjs).

Constructors

File Title Filename Depth [LayoutObj]

type Author = Spec

An author is just a sentence (Spec).

type Contents = Spec

Contents are just a sentence (Spec).

type Items = [LayoutObj]

A group of layout objects.

type Tags = [String]

Tags.

type Depth = Int

Depth of a header.

type Width = Float

Horizontal dimension of a graph.

type Height = Float

Vertical dimension of a graph.

type Filepath = String

Holds a file path.

type Filename = String

Holds a file name.

type Caption = Spec

A caption is just a sentence (Spec).

type RefMap = Map String Filename

A mapping of refs to the file that contains them.

data LayoutObj

Constructors

Table Tags [[Spec]] Label Bool Caption
Holds all information needed for a table.

Header Depth Title Label
Holds all information needed for a header.

Paragraph Contents
Paragraph.

EqnBlock Contents
Equation block.

Definition DType [(String, [LayoutObj])] Label
Definition. Holds the type, contents, and a label.

List ListType
List.

Figure Label (Maybe Caption) Filepath MaxWidthPercent
Holds all information needed for a figure.

Graph [(Spec, Spec)] (Maybe Width) (Maybe Height) Caption Label
Holds all information needed for a graph.

CodeBlock Contents
Code block.

HDiv Tags [LayoutObj] Label
Holds tags, more contents, and a label.

Cell [LayoutObj]

Bib BibRef
Bibliography section.