Language.Drasil.Printing.PrintingInformation

Description

Defines types and functions to gather all the information needed for printing.

Documentation

data Notation

Notation can be scientific or for engineering.

Constructors

Scientific

Engineering

class HasPrintingOptions c where

Able to be printed.

Methods

getSetting :: Lens' c Notation

Holds the printing notation.

newtype PrintingConfiguration

Holds the printing configuration.

Constructors

PC

Fields

_notation :: Notation

notation :: Iso' PrintingConfiguration Notation

data PrintingInformation

Printing information contains a database, a stage, and a printing configuration.

Constructors

PI

Fields

_ckdb :: ChunkDB

_stg :: Stage

_configuration :: PrintingConfiguration

stg :: Lens' PrintingInformation Stage

configuration :: Lens' PrintingInformation PrintingConfiguration

ckdb :: Lens' PrintingInformation ChunkDB

piSys :: System -> Stage -> PrintingConfiguration -> PrintingInformation

Builds a document's printing information based on the system information.

defaultConfiguration :: PrintingConfiguration

Default configuration is for engineering.