Language.Drasil.TeX.Helpers

Description

Defines helper functions used in printing LaTeX documents.

LaTeX Commands

Infrastructre for defining commands, environments, etc. Calls to TP should only occur in this section.

br :: D -> D

Curly braces.

Helper for adding fencing symbols.

sq :: D -> D

Square brackets.

Helper for adding fencing symbols.

parens :: D -> D

Parenthesis.

Helper for adding fencing symbols.

quote :: D -> D

Quotes.

Helper for adding fencing symbols.

lbrace :: D

Helper for opening curly brace.

rbrace :: D

Helper for closing curly brace.

command0 :: String -> D

0-argument command.

command :: String -> String -> D

Make 1-argument command.

commandD :: String -> D -> D

Similar to command, but uses br for braces.

command1o :: String -> Maybe String -> String -> D

1-argument command, with optional argument.

command1oD :: String -> Maybe D -> D -> D

Similar to command1o, but uses sq and br for brackets.

command1p :: String -> String -> String -> D

1-argument command with parameter in square brackets.

command1pD :: String -> D -> D -> D

Similar to command1p, but uses sq and br for brackets.

texSym :: String -> D

Make LaTeX symbol.

command2 :: String -> String -> String -> D

2-argument command.

command2D :: String -> D -> D -> D

Similar to command2, but uses br for brackets.

command3 :: String -> String -> String -> String -> D

3-argument command.

mkEnv :: String -> (D -> D -> D) -> D -> D

Encapsulate environments.

mkEnvArgBr :: String -> String -> D -> D

Encapsulate environments with argument with braces.

mkEnvArgSq :: String -> String -> D -> D

Encapsulate environments with argument with brackets.

mkMinipage :: D -> D

Makes minipage environment.

comm :: String -> String -> Maybe String -> D

For defining (LaTeX) macros.

renewcomm :: String -> String -> D

Renews given command.

empty :: D

Useful to have an empty case.

genSec :: Int -> D

For sections.

ref :: String -> D -> D

For references.

sref :: String -> D -> D

For references.

hyperref :: String -> D -> D

For references.

externalref :: String -> D -> D

For references.

snref :: String -> D -> D

For references.

href :: String -> String -> D

For references.

cite :: String -> Maybe D -> D

For citations.

Define Common LaTeX Commands

count :: String -> D

Newcounter command.

mathbb :: String -> D

Mathbb command.

usepackage :: String -> D

Usepackage command.

includegraphics :: MaxWidthPercent -> String -> D

Include graphics with a given max width percentage.

author :: D -> D

Preamble for a LaTeX document.

caption :: D -> D

Preamble for a LaTeX document.

item :: D -> D

Preamble for a LaTeX document.

label :: D -> D

Preamble for a LaTeX document.

title :: D -> D

Preamble for a LaTeX document.

bold :: D -> D

Preamble for a LaTeX document.

item' :: D -> D -> D

Command for "item".

maketitle :: D

Formatting options for a LaTeX document.

maketoc :: D

Formatting options for a LaTeX document.

newpage :: D

Formatting options for a LaTeX document.

centering :: D

Formatting options for a LaTeX document.

code :: D -> D

Common commands and formatting options for a LaTeX document.

itemize :: D -> D

Common commands and formatting options for a LaTeX document.

enumerate :: D -> D

Common commands and formatting options for a LaTeX document.

description :: D -> D

Common commands and formatting options for a LaTeX document.

description' :: D -> D

Common commands and formatting options for a LaTeX document.

figure :: D -> D

Common commands and formatting options for a LaTeX document.

center :: D -> D

Common commands and formatting options for a LaTeX document.

document :: D -> D

Common commands and formatting options for a LaTeX document.

equation :: D -> D

Common commands and formatting options for a LaTeX document.

symbDescription :: D -> D

Common commands and formatting options for a LaTeX document.

docclass :: String -> String -> D

Command for the document class.

sec :: Int -> D -> D

General section function.

subscript :: D -> D -> D

Makes second argument a subscript of the first argument.

superscript :: D -> D -> D

Makes second argument a superscript of the first argument.

bullet :: D

Macro/Command definitions.

counter :: D

Macro/Command definitions.

ddefnum :: D

Macro/Command definitions.

ddref :: D

Macro/Command definitions.

colAw :: D

Macro/Command definitions.

colBw :: D

Macro/Command definitions.

arrayS :: D

Macro/Command definitions.

modcounter :: D

Macro/Command definitions.

modnum :: D

Macro/Command definitions.

newline :: D -> D

Add newline.

fraction :: D -> D -> D

Create a fraction.

hyperConfig :: D

Configuration settings.

useTikz :: D

Uses luatex85 tex packages.

Helpers

toEqn :: D -> D

toEqn is special; it switches to Math, but inserts an equation environment. Uses resizeExpression macro (defined in Preamble.hs) to prevent page overflow.

paren :: String -> String

Wrap with parenthesis.

Helper(s) for String-Printing in TeX where it varies from HTML/Plaintext.

sqbrac :: String -> String

Wrap with square brackets.

Helper(s) for String-Printing in TeX where it varies from HTML/Plaintext.