Language.Drasil.TeX.Monad

Description

Printing Monad. Starts with a specific data type (reader monad) and extends from there.

Printing Monad

data MathContext

There are two proper contexts, Text and Math; Curr is the current context. There are multiple ways of getting there: for Text, either being at the top-level or inside text. For Math, either surrounded by $ or in begin{equation} .. end{equation}. Curr is when the current context is fine.

Constructors

Text

Math

Curr

newtype PrintLaTeX a

A monad for printing in LaTeX.

Constructors

PL

Fields

runPrint :: MathContext -> a

type D = PrintLaTeX Doc

Convenient abbreviation.

switch :: (MathContext -> MathContext) -> D -> D

MonadReader calls this local. Can switch contexts (including no-switch cases). Adjust printing as necessary.

toMath :: D -> D

Change context to Math.

toText :: D -> D

Change context to Text.

getCtx :: PrintLaTeX MathContext

MonadReader calls this ask.

(%%) :: D -> D -> D infixl 5

Since Text.PrettyPrint steals <>, use %% instead for $$.

($+$) :: D -> D -> D infixr 6

Lifts Text.PrettyPrint's $+$. Above, with no overlapping. Associative.

tpRunPrint :: ([Doc] -> Doc) -> [D] -> D

Concatenates a list of D using a function from [Doc] -> Doc.

vcat :: [D] -> D

List version of $$. Above, except that if the last line of the first argument stops at least one position before the first line of the second begins, these two lines are overlapped.

vpunctuate :: Doc -> [D] -> D

hpunctuate :: Doc -> [D] -> D

nest :: Int -> D -> D

Nest a D by a specified indentation level.

lub :: MathContext -> MathContext -> MathContext

MathContext operations.

newtype Latex

Latex type. Holds Spaces.

Constructors

L

Fields

unPL :: String