Language.Drasil.TeX.Print

Description

Defines main LaTeX printer functions. For more information on each of the helper functions, please view the source files.

Documentation

genTeX :: Document -> PrintingInformation -> Doc

Generates a LaTeX document.

pExpr :: Expr -> D

Print an expression to a document.

pUnit :: USymb -> D

Prints units.

spec :: Spec -> D

Prints all Exprs.

fence :: OpenClose -> Fence -> D

Prints fencing notation ("(),{},|,||").

data OpenClose

Determine wether braces and brackets are opening or closing.

Constructors

Open

Close

pMatrix :: [[Expr]] -> (Doc -> [D] -> D) -> Doc -> (Expr -> D) -> D

For printing a Matrix.

cases :: [(Expr, Expr)] -> (Doc -> [D] -> D) -> Doc -> (Expr -> D) -> D

Helper for printing case expression.