Language.Drasil.DOT.Print

Description

Defines printer types and functions for generating traceability graphs (as .dot files).

Types

type Colour = String

Type synonym for clarity.

type Label = String

Type synonym for clarity.

data NodeFamily

A node family contains a list of UIDs, their display labels, general subgraph label, and colour.

Constructors

NF

Fields

nodeUIDs :: [UID]

Node UIDs.

nodeLabels :: [Label]

Display labels for nodes. We use the reference addresses from the UIDs.

nfLabel :: Label

Individual subgraph labels. These labels do not show on the final generated pdf or png files.

nfColour :: Colour

Gives the ability to change colours of bubbles within the graph.

data GraphInfo

Holds all important and relevant information for generating a traceability graph. Includes nodes, graph edges, and node family information.

Constructors

GI

Fields

assumpNF :: NodeFamily

Assumptions.

ddNF :: NodeFamily

Data definitions.

gdNF :: NodeFamily

General definitions.

tmNF :: NodeFamily

Theory models.

imNF :: NodeFamily

Instance models.

reqNF :: NodeFamily

Requirements (both functional and non-functional).

gsNF :: NodeFamily

Goal statement.

chgNF :: NodeFamily

Changes (both likely and unlikely).

edgesAvsA :: [(UID, [UID])]

Assumptions dependent on assumptions.

edgesAvsAll :: [(UID, [UID])]

Definitions, models, requirements, and changes dependent on assumptions.

edgesRefvsRef :: [(UID, [UID])]

Definitions and models that are dependent on other definitions and models.

edgesAllvsR :: [(UID, [UID])]

Goals and requirements dependent on definitions, models, and other requirements.

edgesAllvsAll :: [(UID, [UID])]

Definitions, models, requirements, goals, and changes that are dependent on one another.

Functions

Main Outputs

outputDot :: FilePath -> GraphInfo -> IO ()

Creates the directory for output, gathers all individual graph output functions and calls them.

mkOutputAvsA :: GraphInfo -> IO ()

Output function for assumptions dependent on assumptions.

mkOutputAvsAll :: GraphInfo -> IO ()

Output function for definitions, models, requirements, and changes dependent on assumptions.

mkOutputRefvsRef :: GraphInfo -> IO ()

Output function for definitions and models that are dependent on other definitions and models.

mkOutputAllvsR :: GraphInfo -> IO ()

Output function for goals and requirements dependent on definitions, models, and other requirements.

mkOutputAllvsAll :: GraphInfo -> IO ()

Output function for definitions, models, requirements, goals, and changes that are dependent on one another.

Helpers

mkOutput :: GraphInfo -> String -> (GraphInfo -> [(UID, [UID])]) -> [GraphInfo -> NodeFamily] -> IO ()

General output function for making a traceability graph. Takes in the graph information, title, edge generator functions, and node family functions.

outputSub :: Handle -> [(UID, [UID])] -> [NodeFamily] -> IO ()

Graph output helper. Takes in the file handle, edges, and node families.

mkDirections :: Handle -> (UID, [UID]) -> IO ()

Prints graph edges (directions) onto a given file handle.

mkNodes :: Handle -> NodeFamily -> IO ()

Prints graph nodes (labels) onto a given file handle.

mkSubgraph :: Handle -> Label -> [UID] -> IO ()

Helper that only makes a subgraph if there are elements in the subgraph. Otherwise, it returns nothing.

filterAndGI :: GraphInfo -> [GraphInfo -> NodeFamily] -> [NodeFamily]

Gets graph labels.

filterInvalidChars :: String -> String

Helper to remove invalid characters.