Language.Drasil.HTML.Helpers

Description

Helper functions for creating HTML printers (specifically, HTML tag wrappers).

Documentation

data BibFormatter

Data type that carries functions that vary for bib printing

Constructors

BibFormatter

Fields

emph :: Doc -> Doc

Emphasis (italics) rendering

spec :: Spec -> Doc

Spec rendering

html :: Doc -> Doc

HTML tag wrapper.

headTag :: Doc -> Doc

Head tag wrapper.

body :: Doc -> Doc

Body tag wrapper.

title :: Doc -> Doc

Title tag wrapper.

paragraph :: Doc -> Doc

Paragraph tag wrapper.

code :: Doc -> Doc

Code tag wrapper.

tr :: Doc -> Doc

Table row tag wrapper.

th :: Doc -> Doc

Table header tag wrapper.

td :: Doc -> Doc

Table cell tag wrapper.

figure :: Doc -> Doc

Figure tag wrapper.

figcaption :: Doc -> Doc

Figcaption tag wrapper.

li :: Doc -> Doc

List tag wrapper.

pa :: Doc -> Doc

Paragraph in list tag wrapper.

ba :: Doc -> Doc

Bring attention to element wrapper.

dd :: Doc -> Doc

Description wrapper

ol :: [String] -> Doc -> Doc

Ordered list tag wrapper.

ul :: [String] -> Doc -> Doc

Unordered list tag wrapper.

table :: [String] -> Doc -> Doc

Table tag wrapper.

dl :: [String] -> Doc -> Doc

Description list wrapper

img :: [(String, Doc)] -> Doc

Image tag wrapper.

h :: Int -> Doc -> Doc

Helper for HTML headers.

data Variation

HTML attribute selector.

Constructors

Class

Id

Align

Title

wrap :: String -> [String] -> Doc -> Doc

General Class wrapper function and formats the document space with cat.

wrap' :: String -> [String] -> Doc -> Doc

General wrapper function and formats the document space with hcat.

wrapGen' :: ([Doc] -> Doc) -> Variation -> String -> Doc -> [String] -> Doc -> Doc

Helper for wrapping HTML tags. The fourth argument provides class names for the CSS.

wrapGen :: Variation -> String -> Doc -> [String] -> Doc -> Doc

General wrapper that formats the document space nicely.

tagL :: String -> Variation -> Doc -> Doc

Helper for creating a left HTML tag with a single attribute.

tagR :: String -> Doc

Helper for creating a right HTML closing tag.

wrapInside :: String -> [(String, Doc)] -> Doc

Helper for wrapping attributes in a tag.

The first argument is tag name.

The Space in the pair is the attribute name,

The Doc is the value for different attributes.

caption :: Doc -> Doc

Helper for setting up captions.

descWrap :: [String] -> Doc -> Doc -> Doc

refwrap' :: String -> Doc -> Doc -> Doc

Helper for wrapping divisions or sections. Arguments: Wrapper element type/tag (e.g., p, div, a), attribute value, body text

refwrap :: Doc -> Doc -> Doc

reflink :: String -> Doc -> Doc

Helper for setting up links to references.

reflinkInfo :: String -> Doc -> Doc -> Doc

Helper for setting up links to references with additional information.

reflinkURI :: String -> Doc -> Doc

Helper for setting up links to external URIs.

image :: Doc -> Maybe Doc -> MaxWidthPercent -> Doc

Helper for setting up figures.

em :: Doc -> Doc

Emphasis (italics) tag.

sup :: Doc -> Doc

Superscript tag.

sub :: Doc -> Doc

Subscript tag.

bold :: Doc -> Doc

Bold tag.

articleTitle :: Doc -> Doc

Title header.

author :: Doc -> Doc

Author header.

divTag :: [String] -> Doc -> Doc

Div tag wrapper.

spanTag :: [String] -> Doc -> Doc

Span tag wrapper.

spanTag' :: Doc -> Doc -> Doc

Span tag wrapper with a title attribute.

indent :: Doc -> Doc

Indent the Document by 2 positions.

makeCases :: [(Expr, Expr)] -> (Expr -> Doc) -> Doc

Create and markup fractions fraction :: Doc -> Doc -> Doc fraction a b = divTag ["fraction"] (spanTag ["fup"] a $$ spanTag ["fdn"] b)

Build case expressions.