Language.Drasil.HTML.Print

Description

Defines all functions needed to print HTML files. For more information on each of the helper functions, please view the source files.

Main Function

genHTML :: PrintingInformation -> String -> Document -> Doc

Generate an HTML document from a Drasil Document.

Citation Renderer

renderCite :: BibFormatter -> Citation -> (Doc, Doc)

For when we add other things to reference like website, newspaper

HTML Bib Formatter

htmlBibFormatter :: BibFormatter

HTML specific bib rendering functions

HTML Spec Printing

pSpec :: Spec -> Doc

Renders the Sentences (Exprs) in the HTML body (called by printLO).

Term Fencing Helpers

data OpenClose

Referring to fence (for parenthesis and brackeds). Either opened or closed.

Constructors

Open

Close

fence :: OpenClose -> Fence -> String

Allows for open/closed variants of parenthesis, curly brackets, absolute value symbols, and normal symbols.