Language.Drasil.Printers

Documentation

data Format

Possible formats for output.

Constructors

TeX

HTML

outputDot :: FilePath -> GraphInfo -> IO ()

Creates the directory for output, gathers all individual graph output functions and calls them.

data GraphInfo

Holds all important and relevant information for generating a traceability graph. Includes nodes, graph edges, and node family information.

Constructors

GI

Fields

assumpNF :: NodeFamily

Assumptions.

ddNF :: NodeFamily

Data definitions.

gdNF :: NodeFamily

General definitions.

tmNF :: NodeFamily

Theory models.

imNF :: NodeFamily

Instance models.

reqNF :: NodeFamily

Requirements (both functional and non-functional).

gsNF :: NodeFamily

Goal statement.

chgNF :: NodeFamily

Changes (both likely and unlikely).

edgesAvsA :: [(UID, [UID])]

Assumptions dependent on assumptions.

edgesAvsAll :: [(UID, [UID])]

Definitions, models, requirements, and changes dependent on assumptions.

edgesRefvsRef :: [(UID, [UID])]

Definitions and models that are dependent on other definitions and models.

edgesAllvsR :: [(UID, [UID])]

Goals and requirements dependent on definitions, models, and other requirements.

edgesAllvsAll :: [(UID, [UID])]

Definitions, models, requirements, goals, and changes that are dependent on one another.

data NodeFamily

A node family contains a list of UIDs, their display labels, general subgraph label, and colour.

Constructors

NF

Fields

nodeUIDs :: [UID]

Node UIDs.

nodeLabels :: [Label]

Display labels for nodes. We use the reference addresses from the UIDs.

nfLabel :: Label

Individual subgraph labels. These labels do not show on the final generated pdf or png files.

nfColour :: Colour

Gives the ability to change colours of bubbles within the graph.

makeCSS :: Document -> Doc

Generates the CSS selectors necessary for a document.

genHTML :: PrintingInformation -> String -> Document -> Doc

Generate an HTML document from a Drasil Document.

makeMd :: [Doc] -> Doc

Combines a list of sentences into a final Doc, also appends end note.

introInfo :: String -> [String] -> Doc

Example title and purpose section.

verInfo :: String -> String -> Doc

Language version section.

unsupOS :: Maybe String -> Doc

Invalid Operating Systems section, does not display unless atleast 1 invalid OS.

regularSec :: Doc -> Doc -> Doc

Constructs regular section section from header and message.

instDoc :: [String] -> Doc

Instruction section, contains 3 paragraphs, Running, Building and Config Files. The Config file section is only displayed if there are configuration files.

extLibSec :: [(String, String)] -> [String] -> Doc

External Libraries section. The inputs are a list of name and version pairs and a list of the corresponding version numbers, these are first combined into a list of triplets, and then each printed on a new line.

endNote :: [String] -> Doc

End section.

data Linearity

Data is either linear or not.

Constructors

Linear

Nonlinear

sentenceDoc :: ChunkDB -> Stage -> Linearity -> Sentence -> Doc

Create sentences for a document in Doc format.

exprDoc :: ChunkDB -> Stage -> Linearity -> Expr -> Doc

Create expressions for a document in Doc format.

codeExprDoc :: ChunkDB -> Stage -> Linearity -> CodeExpr -> Doc

Create code expressions for a document in Doc format.

symbolDoc :: Symbol -> Doc

Create symbols for a document in Doc format.

unitDoc :: Linearity -> USymb -> Doc

Helper for printing units in Doc format.

genTeX :: Document -> PrintingInformation -> Doc

Generates a LaTeX document.

data PrintingInformation

Printing information contains a database, a stage, and a printing configuration.

Constructors

PI

Fields

_ckdb :: ChunkDB

_stg :: Stage

_configuration :: PrintingConfiguration

class HasPrintingOptions c where

Able to be printed.

Methods

getSetting :: Lens' c Notation

Holds the printing notation.

data Notation

Notation can be scientific or for engineering.

Constructors

Scientific

Engineering

defaultConfiguration :: PrintingConfiguration

Default configuration is for engineering.