Language.Drasil.Printers

Formats

data DocType

Document types include Software Requirements Specification and Website. Choosing SRS will generate both TeX and HTML files, while Website generates only as HTML. This also determines what folders the generated files will be placed into.

Constructors

SRS

Website

Lesson

data Format

Possible formats for printer output.

Constructors

TeX

HTML

Jupyter

MDBook

DOT

Types

data GraphInfo

Holds all important and relevant information for generating a traceability graph. Includes nodes, graph edges, and node family information.

Constructors

GI

Fields

assumpNF :: NodeFamily

Assumptions.

ddNF :: NodeFamily

Data definitions.

gdNF :: NodeFamily

General definitions.

tmNF :: NodeFamily

Theory models.

imNF :: NodeFamily

Instance models.

reqNF :: NodeFamily

Requirements (both functional and non-functional).

gsNF :: NodeFamily

Goal statement.

chgNF :: NodeFamily

Changes (both likely and unlikely).

edgesAvsA :: [(UID, [UID])]

Assumptions dependent on assumptions.

edgesAvsAll :: [(UID, [UID])]

Definitions, models, requirements, and changes dependent on assumptions.

edgesRefvsRef :: [(UID, [UID])]

Definitions and models that are dependent on other definitions and models.

edgesAllvsR :: [(UID, [UID])]

Goals and requirements dependent on definitions, models, and other requirements.

edgesAllvsAll :: [(UID, [UID])]

Definitions, models, requirements, goals, and changes that are dependent on one another.

data NodeFamily

A node family contains a list of UIDs, their display labels, general subgraph label, and colour.

Constructors

NF

Fields

nodeUIDs :: [UID]

Node UIDs.

nodeLabels :: [Label]

Display labels for nodes. We use the reference addresses from the UIDs.

nfLabel :: Label

Individual subgraph labels. These labels do not show on the final generated pdf or png files.

nfColour :: Colour

Gives the ability to change colours of bubbles within the graph.

Functions

outputDot :: FilePath -> GraphInfo -> IO ()

Creates the directory for output, gathers all individual graph output functions and calls them.

HTML

Printer

genHTML :: PrintingInformation -> String -> Document -> Doc

Generate an HTML document from a Drasil Document.

Helpers

makeCSS :: Document -> Doc

Generates the CSS selectors necessary for a document.

Markdown

Printer

makeMd :: [Doc] -> Doc

Combines a list of sentences into a final Doc, also appends end note.

Section Printers

introInfo :: String -> [String] -> Maybe String -> Maybe String -> Doc

Example title, authors, and maybe purpose section.

verInfo :: String -> String -> Doc

Language version section.

unsupOS :: Maybe String -> Doc

Invalid Operating Systems section, does not display unless atleast 1 invalid OS.

regularSec :: Doc -> Doc -> Doc

Constructs regular section section from header and message.

instDoc :: [String] -> String -> (String, String) -> Doc

Instruction section, contains 4 paragraphs, Running, Building, Input-Output and Config Files. The Config file section is only displayed if there are configuration files.

extLibSec :: [(String, String)] -> [String] -> Doc

External Libraries section. The inputs are a list of name and version pairs and a list of the corresponding version numbers, these are first combined into a list of triplets, and then each printed on a new line.

endNote :: Int -> [String] -> Doc

End section.

whatInfo :: Maybe String -> Maybe String -> Doc

What section in generated README file, displays description and scope if not empty

Plain

Types

data SingleLine

Data is either linear or not.

Constructors

OneLine

MultiLine

Functions

sentenceDoc :: ChunkDB -> Stage -> SingleLine -> Sentence -> Doc

Create sentences for a document in Doc format.

exprDoc :: ChunkDB -> Stage -> SingleLine -> Expr -> Doc

Create expressions for a document in Doc format.

codeExprDoc :: ChunkDB -> Stage -> SingleLine -> CodeExpr -> Doc

Create code expressions for a document in Doc format.

symbolDoc :: Symbol -> Doc

Create symbols for a document in Doc format.

unitDoc :: SingleLine -> USymb -> Doc

Helper for printing units in Doc format.

showSymb :: Symbol -> String

Helper for printing Symbols

showHasSymbImpl :: HasSymbol x => x -> String

Helper for printing a HasSymbol in Implementation Stage

TeX

genTeX :: Document -> PrintingInformation -> Doc

Generates a LaTeX document.

Jupyter

genJupyter :: PrintingInformation -> DocType -> Document -> Doc

Generate a python notebook document (using json). build : build the SRS document in JSON format build': build the general Jupyter Notbook document

Markdown

genMDBook :: PrintingInformation -> Document -> [(Filename, Doc)]

Generate a mdBook SRS

makeBook :: Document -> PrintingInformation -> Doc

Prints the .toml config file for mdBook.

makeRequirements :: PrintingInformation -> Doc

Prints the .csv file mapping the original filepaths of assets to the location mdBook uses.

Log

printAllDebugInfo :: PrintingInformation -> [Doc]

Gathers all printing functions and creates the debugging tables from them.

Printing Information and Options

data PrintingInformation

Printing information contains a database, a stage, and a printing configuration.

Constructors

PI

Fields

_ckdb :: ChunkDB

_stg :: Stage

_configuration :: PrintingConfiguration

piSys :: System -> Stage -> PrintingConfiguration -> PrintingInformation

Builds a document's printing information based on the system information.

class HasPrintingOptions c where

Able to be printed.

Methods

getSetting :: Lens' c Notation

Holds the printing notation.

data Notation

Notation can be scientific or for engineering.

Constructors

Scientific

Engineering

defaultConfiguration :: PrintingConfiguration

Default configuration is for engineering.